visuals_dataset/requirements.txt
# Waymo dataset (optional - only needed if fetching from GCS)
tensorflow-io-gcs-filesystem
tensorflow
tensorflow-io

# HiKER corruptions backend
Wand>=0.6.7
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0
arcade>=2.6.0
pyvirtualdisplay>=2.0

# General imaging and data processing
Pillow>=9.2.0
numpy>=1.21.5
pyarrow>=10.0.0
pandas>=1.3.0
